refactor(BridgeCrossingStatus) : make new package "bridgestatus" and take BridgeCrossingStatus in
